Support -require for bytecode linking.
This is done in two steps.

First we lookup the cma of `-require`d libraries specified on the
cli. We add these cmas, in order, before (on the left of) all the
other mentioned bytecode objects on the cli and give all that to
Bytelink.link.

In Bytelink.link, when we read the bytecode objects for link
information from left to right and hit a cma, we read its lib_requires
field and resolve these library name to cmas (and recursively) that we
put before (on the left of) the cma in the object list we are scanning.

Once we have this final list of objects we proceed to link as it used
to be the case.