Support -require for native code linking.
This mostly follows what we did for bytecode and again done
in two steps.

First we lookup the cmxa of `-require`d libraries specified on the
cli. We add these cmxas, in order, before (on the left of) all the
other mentioned native code objects on the cli and give all that to
Asmlink.link.

In Asmlink.link, when we read the bytecode objects for link
information from left to right and hit a cmxa, we read its
lib_requires field and resolve these library name to cmxas (and
recursively) that we put before (on the left of) the cmxa in the
object list we are scanning.

Once we have this final list of objects we proceed to link as it used
to be the case.

One twist is that the function [read_objs_infos] is also used by
Asmlink.link_shared, but the latter doesn't resolve the lib_requires
of cmxa files given on the cli, it collects them to add it the
metadata of the final cmxs file (much like Bytelibrarian does). So if
no resolver is provided to [read_objs_infos] it behaves as before.