Asmlink.object_file_name: remove, add obj_info_file_name.
This one needs a bit of care but is a no-op.

All the uses of [object_file_name] is performed on the [objinfos] that
is also given to [read_objs_infos]. All of what gets in the latter
function gets outs with the same file name lookup and erroring
treatment (except in a stringly manner in [object_file_name]),
as such the error paths in [object_file_name] are dead code.

So we make [object_file_name] work on values of type [obj_info] by
turning it into [obj_info_file_name] and apply it on the values
returned by [read_objs_infos].

This shouldn't change anything for now. But will when
[read_objs_infos] starts to add the cmxa files of libraries required
by cmxa files, effectively expanding the original list of [objfiles].