ocaml[nat]: handle directory includes for -require option.
We handle them like we do for OCaml compilation. For each `-require`d
library we add its library directory to the includes in cli order,
after those specified via `-I` on the cli; unless the library is
statically linked in the executable or assumed to be loaded.

Out of the three places where the load path is initialized we only add
them at the last init. In the first [Toploop] toplevel init by side
effect, the cli arguments are not parsed anyways. In the second
[Toploop.set_paths] init which only survives the time for loading
objects on the cli, they are not needed (and will not be needed when
we add the loading of libraries here).

However in the second case we could technically add them if one
considers that calling `ocaml find.cmo -require lib` should allow to
lookup [find.cmo] in the library directory of [lib].