Add `-assume-require LIB` to assume a library was looked up.
In particular for `-linkall` this allows to specify library names in
the executable without having to `-require` them. It is then assumed
that the user specified objects on the cli that correspond to those of
library `LIB`.

Useful if your library is not installed yet or if build systems prevent
compiler library lookups with `-noautoliblink` and specify archives
on the cli themselves.