{T,Opt}opdirs: add load_lib function to API.
This will be used by {T,Opt}opmain. The signature of this function is
largely suboptimal. But we take care not to introduce any new type
dependency like Lib.Name.t because this module is used and visible by
downstream packages while those may not have have access to
`compiler-libs`'s module because of the expunge process. Maybe once
all this will be improved see #7589.